replace int by size_t to avoid overflow

+++ csrc/fp_quantizer/quantize.cpp
@@ -28,7 +28,7 @@ at::Tensor quantize(torch::Tensor& val,
                     int q_bits,
                     int q_mantisa_bits)
 {
-    int total_elems = at::numel(val);
+    size_t total_elems = at::numel(val);
     auto options = at::TensorOptions()
                        .dtype(torch::kInt8)
                        .layout(val.layout())
@@ -39,7 +39,7 @@ at::Tensor quantize(torch::Tensor& val,
                              (q_bits == 6 ? 28.0 :                            // fp6 range
                                   6.0));  // fp4 range (using power 2); TODO (Reza): add the power-4
                                           // in case accuracy is not matching!
-    int num_groups = total_elems / group_size;
+    size_t num_groups = total_elems / group_size;
     auto out = torch::empty({num_groups, group_size * q_bits / 8 + 4}, options);
 
     DISPATCH_QUANTIZE(kHalf, __half, 23, 8);
@@ -68,9 +68,9 @@ void dequantize(torch::Tensor& val,
                 int q_mantisa_bits,
                 int q_exponent_bits)
 {
-    int total_elems = at::numel(val);
+    size_t total_elems = at::numel(val);
 
-    int num_groups = total_elems / group_size;
+    size_t num_groups = total_elems / group_size;
 
     DISPATCH_DEQUANTIZE(kHalf, __half, 10);
 #ifdef BF16_AVAILABLE
@@ -98,9 +98,9 @@ void selective_dequantize(torch::Tensor& val,
                           int q_mantisa_bits,
                           int q_exponent_bits)
 {
-    int total_elems = at::numel(val);
+    size_t total_elems = at::numel(val);
     int num_indexes = indexes.size(0);
-    int num_groups = total_elems / group_size;
+    size_t num_groups = total_elems / group_size;
 
     DISPATCH_DEQUANTIZE_INDEX(kHalf, __half, 10);
 #ifdef BF16_AVAILABLE
